requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
loguru==0.7.2
flask-jwt-extended==4.5.3
celery==5.3.6
redis==5.0.1
